+++ src/types/type.d.ts
@@ -61,32 +61,39 @@ export interface PostDetail extends Post {
 export interface PopularDiaryCardProps {
     genre: string | null;  
     title: string;
-    content: string | null;
+    content: string | null | undefined;
 };
-
-type Bestsellers = {
-  title: string;
-  author: string;
-  cover: string;
-  isbn13?: string;
-};
-
-interface Book {
-  id: string;
-  title: string;
-  description: string;
-  categoryName: string | null;
-}
-
-interface Like {
-  id: string;
-  user_id: string;
-}
 
 interface DiaryPost {
   id: string;
   category: string;
   like: Like[];
   book: Book;
+}
+
+/* 금주의 인기 다이어리 */
+interface Book {
+  id: string;
+  categoryName: string | null;
+  title: string;
+  description: string | null;
+  subInfo?: {
+    subTitle: string;
+  }
+}
+
+interface Like {
+  id: string;
+  user_id: string;
+  reference_category: string;
+  reference_id: string;
+  created_at: string;
+}
+
+interface APIDiaryPost {
+  id: string;
+  category: string;
+  like: Like[];
+  book: Book | null;
 }
